[v1.3.30] refactor legacy components and rename some components
