+++ src/pages/charts/components/MechanicsChart.tsx
@@ -14,7 +14,7 @@ import { useGetMechanicsChartDataMutation } from "../../../services/chartService
 import Strings from "../../../utils/localizations/Strings";
 import CustomLegend from "../../../components/CustomLegend";
 import { useSearchCardsQuery } from "../../../services/cardService";
-import CustomDrawerCardList from "../../../components/CustomDrawerCardList";
+import DrawerTagList from "../../components/DrawerTagList";
 
 export interface MechanicsChartProps {
   siteId: string;
@@ -130,7 +130,7 @@ const MechanicsChart = ({
           ))}
         </BarChart>
       </ResponsiveContainer>
-      <CustomDrawerCardList
+      <DrawerTagList
         open={open}
         dataSource={searchData}
         isLoading={isFetching}
