+++ src/pages/charts/components/MethodologiesChart.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { useSearchCardsQuery } from "../../../services/cardService";
-import CustomDrawerCardList from "../../../components/CustomDrawerCardList";
+import DrawerTagList from "../../components/DrawerTagList";
 import Strings from "../../../utils/localizations/Strings";
 import {
   Cell,
@@ -125,8 +125,7 @@ const MethodologiesChart = ({ siteId, methodologies }: MethodologiesChartProps) 
           />
         </PieChart>
       </ResponsiveContainer>
-      {/* Componente para mostrar la lista de cards/tags */}
-      <CustomDrawerCardList
+      <DrawerTagList
         open={open}
         dataSource={searchData}
         isLoading={isFetching}
